Reorganized forms and cleaned up includes
