+++ CMakeLists.txt
@@ -5,8 +5,6 @@ set(CMAKE_CXX_STANDARD 17)
 set (CMAKE_EXPORT_COMPILE_COMMANDS ON)
 
 include_directories(.)
-include_directories(impl)
-include_directories(mavis)
 
 add_library(mavis
   impl/ExtractorRegistry.cpp
