+++ mavis/FormRegistry.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "FormIF.h"
+#include "Form.h"
 #include <string>
 
 namespace mavis {
