+++ mavis/InstMetaData.h
@@ -2,7 +2,7 @@
 
 #include "json.hpp"
 #include "DecoderExceptions.h"
-#include "FormIF.h"
+#include "Form.h"
 #include "Tag.hpp"
 #include "MatchSet.hpp"
 #include <memory>
